Refactor : moved productItem to seperate components
