+++ src/Components/Products.js
@@ -1,11 +1,12 @@
-import { useEffect, useReducer } from "react";
+import { useEffect, useReducer, useState } from "react";
 import { useData } from "../Context";
 import { Actions } from "../Context";
 import axios from "axios";
-// import { useAxios } from "./useAxios";
+import { ProductItem } from "./ProductItem.jsx";
 
 export const Products = () => {
   const { state, dispatch } = useData();
+  const [loadError, setLoadError] = useState(false);
   useEffect(() => {
     document.title = "Products";
   }, []);
@@ -18,7 +19,7 @@ export const Products = () => {
           payload: products.products,
         });
       } catch (err) {
-        console.log(err);
+        setLoadError(true);
       }
     })();
   }, [dispatch]);
@@ -94,38 +95,41 @@ export const Products = () => {
 
   return (
     <div>
+      {loadError && <h2>Error in loading Data</h2>}
       {state.products.length > 0 ? (
         <div>
           <div>
-            <span className="text">Sort using price</span>
-            <label>
-              <input
-                type="radio"
-                name="sort"
-                onChange={() =>
-                  filterDispatch({
-                    type: "SORTBY",
-                    payload: "SORT_LOW_TO_HIGH",
-                  })
-                }
-                checked={sortBy && sortBy === "SORT_LOW_TO_HIGH"}
-              />
-              LOW TO HIGH
-            </label>
-            <label>
-              <input
-                type="radio"
-                name="sort"
-                onChange={() =>
-                  filterDispatch({
-                    type: "SORTBY",
-                    payload: "SORT_HIGH_TO_LOW",
-                  })
-                }
-                checked={sortBy && sortBy === "SORT_HIGH_TO_LOW"}
-              />
-              HIGH TO LOW
-            </label>
+            <span className="text">Sort using price :</span>
+            <div>
+              <label>
+                <input
+                  type="radio"
+                  name="sort"
+                  onChange={() =>
+                    filterDispatch({
+                      type: "SORTBY",
+                      payload: "SORT_LOW_TO_HIGH",
+                    })
+                  }
+                  checked={sortBy && sortBy === "SORT_LOW_TO_HIGH"}
+                />
+                LOW TO HIGH
+              </label>
+              <label>
+                <input
+                  type="radio"
+                  name="sort"
+                  onChange={() =>
+                    filterDispatch({
+                      type: "SORTBY",
+                      payload: "SORT_HIGH_TO_LOW",
+                    })
+                  }
+                  checked={sortBy && sortBy === "SORT_HIGH_TO_LOW"}
+                />
+                HIGH TO LOW
+              </label>
+            </div>
           </div>
           <div>
             <span className="text">Filters : </span>
@@ -165,7 +169,7 @@ export const Products = () => {
           <ul className="products-list">
             {searchedData.length > 0 ? (
               searchedData.map((item) => (
-                <Product key={item.id} product={item} />
+                <ProductItem key={item.id} product={item} />
               ))
             ) : (
               <h2 className="text text-large">
@@ -177,84 +181,6 @@ export const Products = () => {
       ) : (
         <h2>Loading Products</h2>
       )}
-    </div>
-  );
-};
-
-const Product = ({ product }) => {
-  const { dispatch, setToast, setToastMessage, state } = useData();
-  const inCartProducts = (id) => {
-    return state.cartProducts.findIndex((item) => item.id === id) === -1
-      ? false
-      : true;
-  };
-  const inWishProducts = (id) => {
-    return state.wishProducts.findIndex((item) => item.id === id) === -1
-      ? false
-      : true;
-  };
-
-  return (
-    <div className="product">
-      <div className="card card-shopping">
-        <img src={product.url} alt="" className="card-img" />
-        <h4 className="card-brand">{product.brandName}</h4>
-        <p className="card-desc">{product.description}</p>
-        <p className="card-desc">Rs {product.price}</p>
-
-        {inWishProducts(product.id) ? (
-          <span
-            onClick={() => {
-              setToast("true");
-              setToastMessage(`${product.brandName} removed from  wishlist`);
-              dispatch({
-                type: Actions.REMOVE_FROM_WISHLIST,
-                payload: product.id,
-              });
-            }}
-            className="wish-button"
-            role="button"
-          >
-            <i class="fas fa-heart"></i>
-          </span>
-        ) : (
-          <span
-            onClick={() => {
-              setToast("true");
-              setToastMessage(`${product.brandName} is added to wishlist`);
-              dispatch({ type: Actions.ADD_TO_WISHLIST, payload: product.id });
-            }}
-            className="wish-button"
-            role="button"
-          >
-            <i class="far fa-heart"></i>
-          </span>
-        )}
-
-        {inCartProducts(product.id) ? (
-          <button
-            onClick={() => {
-              setToast("true");
-              setToastMessage(`${product.brandName} is added to Cart`);
-              dispatch({ type: Actions.REMOVE_FROM_CART, payload: product.id });
-            }}
-            className="button button-primary"
-          >
-            REMOVE FROM CART
-          </button>
-        ) : (
-          <button
-            onClick={() => {
-              setToast("true");
-              setToastMessage(`${product.brandName} is added to Cart`);
-              dispatch({ type: Actions.ADD_TO_CART, payload: product.id });
-            }}
-            className="button button-primary"
-          >
-            ADD TO CART
-          </button>
-        )}
-      </div>
     </div>
   );
 };
