+++ src/Context/ActionTypes.js
@@ -9,3 +9,5 @@ export const DECREASE_ITEM_IN_WISHLIST = "DECREASE_ITEM_IN_WISHLIST";
 export const INCREASE_ITEM_IN_CART = "INCREASE_ITEM_IN_CART";
 export const DECREASE_ITEM_IN_CART = "DECREASE_ITEM_IN_CART";
 export const SET_PRODUCTS_DATA = "SET_PRODUCTS_DATA";
+export const SET_WISHLIST_DATA = "SET_WISHLIST_DATA";
+export const SET_CART_DATA = "SET_CART_DATA";
